backend/services/xiaomi_service.py
import logging
from miio import Device
from miio.device import DeviceException
import time
from datetime import datetime
import os

logger = logging.getLogger(__name__)

class XiaomiScaleService:
    def __init__(self, token=None, ip=None):
        """
        Initialize the Xiaomi Scale service with device token and IP.
        
        Args:
            token (str): The token for the Xiaomi device
            ip (str): The IP address of the Xiaomi device
        """
        self.token = token or os.environ.get('XIAOMI_TOKEN')
        self.ip = ip or os.environ.get('XIAOMI_IP')
        self.device = None
        
    def connect(self):
        """
        Connect to the Xiaomi device.
        
        Returns:
            bool: True if connection successful, False otherwise
        """
        if not self.token or not self.ip:
            logger.error("Xiaomi device token or IP not provided")
            return False
            
        try:
            self.device = Device(ip=self.ip, token=self.token)
            info = self.device.info()
            logger.info(f"Connected to Xiaomi device: {info.model}")
            return True
        except DeviceException as e:
            logger.error(f"Failed to connect to Xiaomi device: {e}")
            return False
            
    def get_scale_data(self):
        """
        Get data from the Xiaomi Scale.
        
        Returns:
            dict: The scale data including weight, body fat, etc.
        """
        if not self.device:
            if not self.connect():
                return None
                
        try:
            # Note: The exact command may vary depending on your Xiaomi Scale model
            # For demonstration purposes - in real implementation, you'd use the
            # correct method for your specific device model
            data = self.device.send("get_weight_data")
            
            # Process and format the data
            processed_data = {
                'weight': data.get('weight', 0) / 1000,  # Convert g to kg
                'bmi': data.get('bmi', 0) / 10,  # Scale factor
                'body_fat': data.get('bodyfat', 0) / 10,  # Percentage
                'muscle_mass': data.get('muscle', 0) / 1000,  # Convert g to kg
                'water': data.get('water', 0) / 10,  # Percentage
                'visceral_fat': data.get('visceral', 0),
                'bone_mass': data.get('bone', 0) / 1000,  # Convert g to kg
                'basal_metabolism': data.get('basal', 0),  # kcal
                'protein': data.get('protein', 0) / 10,  # Percentage
                'timestamp': datetime.utcnow().isoformat(),
                'source': 'xiaomi'
            }
            
            return processed_data
        except DeviceException as e:
            logger.error(f"Error retrieving data from Xiaomi device: {e}")
            return None
    
    def discover_devices(self):
        """
        Discover Xiaomi devices on the network.
        
        Returns:
            list: List of discovered devices
        """
        try:
            # This is a simplified example - actual discovery depends on the library implementation
            from miio.discovery import discover
            
            devices = discover()
            return [
                {
                    'ip': device.ip,
                    'id': device.did,
                    'model': device.model
                }
                for device in devices
            ]
        except Exception as e:
            logger.error(f"Error discovering Xiaomi devices: {e}")
            return [] 