backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.19.1
Werkzeug==2.0.1
marshmallow==3.14.1
python-miio==0.5.12
pandas==1.3.4
numpy==1.21.4
scikit-learn==1.0.1
statsmodels==0.13.1
pymongo==4.0.1
matplotlib==3.5.0
seaborn==0.11.2
gunicorn==20.1.0
pytest==6.2.5 